+++ src/Dekatrian.php
@@ -24,26 +24,29 @@
  * @see       https://github.com/vitorteccom/dekaph Repository of Dekaph
  */
 
+declare(strict_types=1);
+
 namespace Vitorteccom\Dekaph;
+
+use BadMethodCallException;
 
 /**
  * Some methods to convert dates between Dekatrian and Gregorian calendars
  */
 class Dekatrian
 {
-
     /**
      * Check if a Dekatrian date is valid or not
      *
-     * @param integer $year  Dekatrian year
-     * @param integer $month Dekatrian month
-     * @param integer $day   Dekatrian day
+     * @param int $year  Dekatrian year
+     * @param int $month Dekatrian month
+     * @param int $day   Dekatrian day
      *
-     * @return boolean Dekatrian date is valid or not
+     * @return bool Dekatrian date is valid or not
      */
-    private static function checkDekatrian($year, $month, $day)
+    private static function checkDekatrian(int $year, int $month, int $day): bool
     {
-        if (date("L", mktime(0, 0, 0, 1, 1, $year))) {
+        if (date('L', mktime(0, 0, 0, 1, 1, $year))) {
             if ($month == 0 && $day > 2) {
                 return false;
             }
@@ -63,68 +66,72 @@ class Dekatrian
     /**
      * Convert a Dekatrian date to Gregorian date
      *
-     * @param integer $year  Dekatrian year
-     * @param integer $month Dekatrian month
-     * @param integer $day   Dekatrian day
+     * @param int $year  Dekatrian year
+     * @param int $month Dekatrian month
+     * @param int $day   Dekatrian day
+     *
+     * @throws BadMethodCallException   If not a valid Dekatrian date
      *
      * @return string Gregorian date, e.g. 2018-12-31
      */
-    public static function dekaToGreg($year, $month, $day)
+    public static function dekaToGreg(int $year, int $month, int $day): string
     {
         if (!self::checkDekatrian($year, $month, $day)) {
-            return false;
+            throw new BadMethodCallException('The provider arguments are not a valid Dekatrian date');
         }
 
         $yearDay = ($month*28)-28;
         $yearDay = $yearDay+$day;
 
-        // leap year
-        if (date("L", mktime(0, 0, 0, 1, 1, $year))) {
+        /** leap year */
+        if (date('L', mktime(0, 0, 0, 1, 1, $year))) {
             $yearDay++;
 
-            // Sinchronian day
+            /** Sinchronian day */
             if ($month == 0 && $day == 2) {
-                return $year ."-1-2";
+                return $year .'-1-2';
             }
         }
 
-        //Achronian day
+        /** Achronian day */
         if ($month == 0 && $day == 1) {
-            return $year ."-1-1";
+            return $year .'-1-1';
         }
 
         $mktime = mktime(0, 0, 0, 1, 1, $year);
-        return date("Y-m-d", strtotime("+$yearDay day", $mktime));
+        return date('Y-m-d', strtotime("+$yearDay day", $mktime));
     }
 
     /**
      * Convert a Gregorian date to Dekatrian date
      *
-     * @param integer $year  Gregorian year
-     * @param integer $month Gregorian month
-     * @param integer $day   Gregorian day
+     * @param int $year  Gregorian year
+     * @param int $month Gregorian month
+     * @param int $day   Gregorian day
+     *
+     * @throws BadMethodCallException   If not a valid Gregorian date
      *
      * @return string Dekatrian date, e.g. 2018-13-28
      */
-    public static function gregToDeka($year, $month, $day)
+    public static function gregToDeka(int $year, int $month, int $day): string
     {
         if (!checkdate($month, $day, $year)) {
-            return false;
+            throw new BadMethodCallException('The provider arguments are not a valid Gregorian date');
         }
 
-        $yearDay = date("z", mktime(0, 0, 0, $month, $day, $year));
+        $yearDay = date('z', mktime(0, 0, 0, $month, $day, $year));
 
-        // leap year
-        if (date("L", mktime(0, 0, 0, $month, $day, $year))) {
+        /** leap year */
+        if (date('L', mktime(0, 0, 0, $month, $day, $year))) {
             $yearDay--;
 
-            // Sinchronian day
+            /** Sinchronian day */
             if ($month == 1 && $day == 2) {
                 return $year .'-0-2';
             }
         }
 
-        // Achronian day
+        /** Achronian day */
         if ($month == 1 && $day == 1) {
             return $year .'-0-1';
         }