+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## [3.1.0] - 2019-01-29
+
+### Added
+- support Exception and strict_types
+
 ## [3.0.1] - 2018-12-22
 
 ### Changed
@@ -13,7 +18,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ### Changed
 
-- dekatoGreg method to dekaToGreg 
+- dekatoGreg method to dekaToGreg
 
 ## [2.0.1] 2018-07-18